+++ CHANGELOG.md
@@ -69,6 +69,7 @@
 - invalid string in ast dump
 - correctly propagate unreachable code error from loaded module
 - incorrect line number after call ``DSState_loadModule`` api
+- not perform file name completion after ``:``
 
 ## [0.26.0] - 2022-06-30